pyproject.toml
[project]
name = "python-drone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiortc>=1.14.0",
    "bleak>=1.1.1",
    "opencv-python>=4.12.0.88",
    "paho-mqtt>=2.1.0",
    "websockets>=15.0.1",
]
